requirements.txt
pandas
pdfplumber
Pillow
pyzotero
Requests
services
streamlit
streamlit_option_menu
psycopg2-binary
python-dotenv
streamlit-authenticator
